Extracted remaining commands
